gio/gdbusaddress.c: Fix RunDLL call for Windows/x64 builds
On x64 builds under both Visual C++ and MinGW-64, the correct symbol for
the RunDLL call is g_win32_run_session_bus, so this will silence the
annoying RunDLL error for x64 MinGW builds as well, and will allow this
call to be executed more than once.

Thanks to Tarnyko for pointing this out.