Fix and enhance GTimeZone for Windows
Use RegLoadMUIStringW() to query MUI_Std and MUI_Dlt when possible,
as this has the advantage to retrieve a specific string for the timezone
name and abbreviation in a specific language, rather than being
hard-connected to the language version of Windows where the program is
being run on (i.e. the Std and Dlt entries for a given time zone name).

Also fix rule_from_windows_time_zone_info() as we can't just do a strncpy()
of tzi->StandardName and tzi->DaylightName directly, as they are wchar_t/
gunichar2 strings, and use the wide variants of RegOpenKeyEx()/
RegQueryValueEx() throughout so that we acquire tzi->StandardName and
tzi->DaylightName consistently on all the supported versions of Windows.

https://bugzilla.gnome.org/show_bug.cgi?id=719344