docs: Add a README.rationale documenting major design decisions
It will be useful to document the major decisions which affect the whole
of GLib in one centralised, easily-greppable file, otherwise they will
get lost forever in Bugzilla.

This file should contain a brief explanation of the decision and its
rationale, plus a link to further discussion (e.g. on a mailing list or
bug report).

This contains an initial discussion about use of compiler attributes in
GLib.

https://bugzilla.gnome.org/show_bug.cgi?id=113075