Update config.h.win32.in for Newer Windows
Make use of if_indextoname() and if_nametoindex() when building against
Window Vista/Server 2008 or later, as these are provided by the system.

This is not turned on by default as we still want to support XP and
Server 2003-turn this on by changing _WIN32_WINNT to 0x600 or later prior
to compiling GLib.

https://bugzilla.gnome.org/show_bug.cgi?id=730352